change code struct
